+++ components/RacialDiversityGraph.jsx
@@ -1,26 +1,22 @@
 import { ResponsiveLine } from '@nivo/line';
-import heroRatio from '../constants/heroRatio.json';
+import racialDiversity from '../constants/is390/racialDiversity.json';
 
-export default function HeroRatioGraph() {
+export default function RacialDiversityGraph() {
   return (
     <ResponsiveLine
-      markers={[
-        {
-          axis: 'y',
-          value: 0.5,
-          legend: '50%'
-        },
-      ]}
-      colors={{ scheme: 'category10' }}
+      colors={{ scheme: 'dark2' }}
       enableSlices="x"
-      data={heroRatio}
+      data={racialDiversity}
       isInteractive
+      useMesh
       animate
-      margin={{ top: 50, right: 130, bottom: 50, left: 80 }}
+      margin={{ top: 50, right: 110, bottom: 50, left: 80 }}
       xScale={{ type: 'point' }}
       yScale={{
         type: 'linear',
-        stacked: false,
+        min: 'auto',
+        max: 'auto',
+        stacked: true,
         reverse: false,
       }}
       yFormat=" >-.5f"
@@ -40,12 +36,14 @@ export default function HeroRatioGraph() {
         tickSize: 5,
         tickPadding: 5,
         tickRotation: 0,
-        legend: 'ratio of good to total characters',
+        legend: 'racial diversity score',
         legendOffset: -55,
         legendPosition: 'middle',
       }}
       pointSize={10}
+      pointColor={{ theme: 'background' }}
       pointBorderWidth={2}
+      pointBorderColor={{ from: 'serieColor' }}
       pointLabelYOffset={-12}
       useMesh={true}
       legends={[
