+++ constants/is390/write.js
