+++ package.json
@@ -20,10 +20,16 @@
     "framer-motion": "^4.1.17",
     "next": "^9.5.4",
     "node-fetch": "^2.6.1",
+    "object-path": "^0.11.5",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-icons": "^4.1.0",
     "react-player": "^2.7.2",
     "react-scroll": "^1.8.1"
+  },
+  "devDependencies": {
+    "@types/node": "^16.3.3",
+    "@types/react": "^17.0.14",
+    "typescript": "^4.3.5"
   }
 }
